First step towards Fig Standards - no functional changes
This massive commit reformats the files to use proper spaces not tabs.
In addition it changes the case on certain keywords to adhere to standards - downcast of AND, OR, FALSE, TRUE and NULL
Bracket {} alignment is modified to meet standards
Proper indentation of multiple lines and some line wrapping (incomplete) are accomplished.