Fix unclear reference that should have been to function secure()
This module has a protected variable which does not meet Fig Standards as it uses a leading "_" to indicate visually it is not public. In addition there is a function secure with an optional parameter. The reference in the calling method used neither, referencing a variable using magic method. This means it was not using the protected value nor calling the function and would always test false as it was NULL. That is a bug.