Handle empty API key
The existing approach would only prevent reporting if the configured API
key was explicitly of type `null`. If the key was an empty string, the
API call would be made, and fail.

It is common for API keys to be defined in a `.env` file, and also
common for a values to be defined empty, especially in development
environments when they are copied from a `.env.example` file and not
being used for local development. When a value is left empty, it is
parsed as an empty string rather than a null.