Baby step towards guzzle client config override
I was having issues with the guzzle "verify" client option. By default HoneybaderClient sets "verify" to true. Sometimes this is NOT the best option because some developers do NOT have access to the php.ini or openssl configuration. https://docs.guzzlephp.org/en/stable/request-options.html#verify

I commented on #112 (https://github.com/honeybadger-io/honeybadger-php/issues/112#issuecomment-777080461) and @joshuap provided that another issue #121 may be related.

Took a little time and found that in Laravel the client configuration is not brought in generally, instead is specified with specific context as seen in lines 91-92.

I've added the "verify" option that works from Laravel's honeybader configuration. Easy test is to set "verify" to false.